Bug fix... Invalid shift value in prefixCoded bytes, closes #5

+++ maalr.lucene/src/main/java/de/uni_koeln/spinfo/maalr/lucene/core/Dictionary.java
@@ -253,7 +253,7 @@ public class Dictionary {
 			query = bc;
 			TopDocs docs = indexProvider.getSearcher().search(query,
 					null, pageSize,
-					new Sort(new SortField(sortField, SortField.Type.INT)));
+					new Sort(new SortField(sortField, SortField.Type.LONG)));
 
 			return toQueryResult(docs, 0, pageSize);
 		} catch (IOException e) {
